Refactor bsc package and CS leftovers
